Adding the .APNG extension
This should add recognition of .apng files to the sticker manager, which some animated stickers are formated in instead of .gif. GBoard and associated apps "should" natively support it, as it is the preferred sticker format of messaging apps like LINE, but right now I have no way to check and confirm it. Love the app!